+++ NEWS.md
@@ -10,6 +10,7 @@
   - Increases speed
   - Columns have better types (e.g., year, month, day all numeric now)
   - Column order now reflects original data
+- Add option to download only specific months (#68)
 
 
 # weathercan 0.7.8